+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # HEAD
 
+# HealpixMPI v1.0.0
+
+**Merged pull requests:**
+- Add spin-2 sht (#4) (@LeeoBianchi)
+
+**What's new:**
+- Polarization support for SHT
+- Polarization support for distributed datatypes DAlm & DMap
+- Polarization support for MPI.Scatter!, MPI.Gather! and MPI.Allgather! overload methods, to allow direct portability to and from Healpix.jl
+- DAlm & DMap have now abstract super-types
+
+**Braking changes:**
+- Removed the Integer type specification in the signature of AlmInfoMPI and GeomInfoMPI, and of DAlm and DMap consequently. 
+
 # Version 0.1.0
 
 **Merged pull requests:**